refactor(pattern): split tests into unit and integration
Move 9 public API tests (scan_entities, threshold, boost, pattern
selection, dictionary confidence) to tests/scan.rs as integration
tests. Keep 11 unit tests that need pub(crate) access (scan_raw,
allow/deny lists, builder internals, into_entity) in engine/mod.rs.
